refactor(wordle): Extract client-side API functions to a separate module
