+++ apps/wordle/src/frontend/discordApi.ts
@@ -1,0 +1,62 @@
+import type { DiscordSDK } from "@discord/embedded-app-sdk";
+import { REST } from "@discordjs/rest";
+import {
+  CDNRoutes,
+  ImageFormat,
+  type ImageSize,
+  type RESTGetAPICurrentUserGuildsResult,
+  RouteBases,
+  Routes,
+} from "discord-api-types/v10";
+
+export async function getAuthorizationCode(sdk: DiscordSDK): Promise<string> {
+  const { code } = await sdk.commands.authorize({
+    client_id: sdk.clientId,
+    response_type: "code",
+    scope: ["identify", "guilds"],
+  });
+  return code;
+}
+
+export async function getAccessToken(code: string): Promise<string> {
+  const tokenRes = await fetch("/.proxy/api/token", {
+    method: "POST",
+    headers: {
+      "Content-Type": "application/json",
+    },
+    body: JSON.stringify({ code }),
+  });
+
+  const { access_token } = (await tokenRes.json()) as { access_token: string };
+  if (typeof access_token !== "string") {
+    throw new Error("Invalid access token received");
+  }
+
+  return access_token;
+}
+
+export async function fetchGuilds(accessToken: string) {
+  return getRESTClient(accessToken).get(
+    Routes.userGuilds(),
+  ) as Promise<RESTGetAPICurrentUserGuildsResult>;
+}
+
+export function getGuildIconUrl(
+  guildId: string,
+  iconHash: string,
+  size: ImageSize,
+): string {
+  const route = CDNRoutes.guildIcon(guildId, iconHash, ImageFormat.PNG);
+  const url = new URL(route, RouteBases.cdn);
+  return `${url}?${getCDNQueryParams(size)}`;
+}
+
+export function getRESTClient(accessToken?: string) {
+  const rest = new REST({ authPrefix: "Bearer" });
+  if (accessToken) rest.setToken(accessToken);
+  return rest;
+}
+
+export function getCDNQueryParams(size: ImageSize) {
+  return new URLSearchParams({ size: size.toString() });
+}
